requirements.txt
openai==1.68.2
azure-ai-formrecognizer==3.2.1
azure-core==1.29.5
python-dotenv==1.0.1
fastapi==0.110.2
uvicorn==0.29.0
pandas==2.1.4
PyMuPDF==1.23.3  # For PDF parsing
streamlit
azure-identity
azure-storage-blob
